requirements.txt
streamlit
tensorflow
scikit-learn
pandas
numpy
joblib
matplotlib
opencv-python-headless
Pillow